+++ packages/polaris-viz/src/components/SparkBarChart/tests/SparkBarChart.test.tsx
@@ -5,6 +5,7 @@ import {LinearGradientWithStops} from '@shopify/polaris-viz-core';
 import type {DataSeries} from '@shopify/polaris-viz-core';
 
 import {SparkBarChart} from '../SparkBarChart';
+import {Chart} from '../Chart';
 
 const sampleData: DataSeries = {
   data: [
@@ -132,6 +133,26 @@ describe('<SparkBarChart/>', () => {
     expect(wrapper).toContainReactComponent('path', {
       strokeDasharray: '18.5 11.5',
       strokeDashoffset: -25.75,
+    });
+  });
+
+  describe('theme', () => {
+    it('gets passed to <Chart/>', () => {
+      const sparklineChart = mount(
+        <SparkBarChart data={[sampleData]} theme="Light" />,
+      );
+
+      expect(sparklineChart).toContainReactComponent(Chart, {
+        theme: 'Light',
+      });
+    });
+
+    it('`Default` is used if no value is provided', () => {
+      const barChart = mount(<SparkBarChart data={[sampleData]} />);
+
+      expect(barChart).toContainReactComponent(Chart, {
+        theme: 'Default',
+      });
     });
   });
 });