+++ packages/polaris-viz-core/CHANGELOG.md
@@ -11,6 +11,13 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Updated react-spring to v9.4.5
 
+## [1.10.0] - 2022-05-25
+
+### Added
+
+- `ChartProps` and `WithRequired` types
+- `DEFAULT_CHART_PROPS` constant
+
 ## [1.9.3] - 2022-05-25
 
 - No updates. Transitive dependency bump.