+++ packages/polaris-viz/CHANGELOG.md
@@ -11,6 +11,12 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Updated react-spring to v9.4.5
 
+## [1.10.0] - 2022-05-25
+
+### Changed
+
+- `ChartProps` and `DEFAULT_CHART_PROPS` are now used on `<LineChart/>`, `<BarChart/>`, `<SimpleBarChart/>`, `<SparkBarChart/>`, `<SparklineChart/>` and `<StackedAreaChart/>`
+
 ## [1.9.3] - 2022-05-25
 
 ### Fixed